update: HeadBar等のコンポーネントを再分割
